requirements.txt
google-generativeai>=0.8.2
tweepy>=4.14.0
apify-client>=1.7.1
python-dotenv>=1.0.1
